requirements-dev.txt
pytest>=8.3.0
mypy>=1.11.0
black>=24.4.0
ruff>=0.6.0
pre-commit>=3.7.0
hypothesis>=6.95.0
jsonschema>=4.23.0

# ... other dev tools like pytest, black, etc.
mdformat
PyYAML>=6.0.0
